+++ Changelog.md
@@ -7,6 +7,7 @@ All notable changes to top-ii-vol mesher will be documented in this file.
 
 ### Removed
 - Merged the partitioner and distributed mesher, no more `top-ii-vol-Part`
+- No more `--scatter`, now a part of parallel mesher.
 
 ## [1.1] 2020‑03‑31
 ### Added